+++ trunk/ContraGame/ContraGame/CMBullet.cpp
@@ -19,7 +19,7 @@ void CMBullet::LoadResources()
 	_current_sprite = _bullet_sprite;
 }
 
-void CMBullet::Update(float delta_time)
+void CMBullet::Update(int delta_time)
 {
 	_physical.CalcPositionWithoutGravitation(delta_time);
 	_physical.SetBounds(
