+++ trunk/ContraGame/ContraGame/CNEBullet.h
@@ -13,7 +13,7 @@ public:
 	~CNEBullet();
 
 	void LoadResources();
-	void Update(float delta_time);
+	void Update(int delta_time);
 	void Draw();
 
 	void Moving(float v_max);
